+++ crates/rust-releases-bundled/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Added
 
-- Add offline `BundledReleases` "client", which bundles the stable, beta and nightly releases of Rust as generated Rust code
+- Add crate which bundles the stable, beta and nightly releases of Rust as generated Rust code